Seperate Catch2 and Unity includes and runners
There are some macros which Unity and Catch2 define with the same name
and there were compilation errors so the Catch2 initializer function has
been moved to another file.
Also the include path has been changed for Catch2 to include the header
file within the project and to prevent from including the system wide
header which could be installed using a package manager.