Rewrite sequence_finder tests to use GENERATE
This 3rd generation of the test case uses GENERATE(), which is Catch's
keyword used to create parametrized test cases. At this stage the DRY
rule applies to the tests, but the tests are really hard to follow.
Custom failure messages are applied and no REQUIRE/CHECK macros are used
what kind'a breaks the TDD approach.
Maybe it shall be done like that, that each test shall be unrolled, and
each line shall contain a single:

REQUIRE(sequence_finder.found('c'));

?